Added frontend stuff and did things with floats in the db and models

+++ Music_Review_Application_Database/dbo/Tables/AlbumReview.sql
@@ -3,7 +3,7 @@
   [id] int PRIMARY KEY IDENTITY,
   [albumId] int not null,
   [username] nvarchar(30) not null,
-  [albumScore] int not null,
+  [albumScore] float not null,
   [albumReview] nvarchar(2500), 
     CONSTRAINT [FK_AlbumReview_ToAlbumTable] FOREIGN KEY ([albumId]) REFERENCES [Album]([id])
 )
